fix(workspaces): restore Copy action from the workspace list
Choosing Copy from a workspace's three-dots menu threw
"this.toCopyWorkspace is not a function". The action lived on the workspaces
route, but the list template is rendered by the workspaces/index controller
(via templateName), so {{this.toCopyWorkspace}} never resolved. Move it to the
controller.

Files:
- app/controllers/workspaces/index.js  (new: toCopyWorkspace -> workspaces.copy)
- app/routes/workspaces.js             (drop the unreachable action)
